pyproject.toml
[project]
name = "voice-control-agent"
version = "0.1.0"
description = "Voice-controlled agent for computer control and Claude Code integration"
requires-python = ">=3.10"
dependencies = [
    "pipecat-ai[google,deepgram,silero,webrtc,runner]>=0.0.99",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff",
    "ty",
]

[tool.ty.environment]
python-version = "3.12"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
extend-select = [
    "F",    # Pyflakes rules
    "W",    # PyCodeStyle warnings
    "E",    # PyCodeStyle errors
    "I",    # Sort imports properly
    "UP",   # Warn if certain things can be changed due to newer Python versions
    "C4",   # Catch incorrect use of comprehensions, dict, list, etc
    "FA",   # Enforce from __future__ import annotations
    "ISC",  # Good use of string concatenation
    "ICN",  # Use common import conventions
    "RET",  # Good return practices
    "SIM",  # Common simplification rules
    "TID",  # Some good import practices
    "TC",   # Enforce importing certain types in a TYPE_CHECKING block
    "PTH",  # Use pathlib instead of os.path
    "TD",   # Be diligent with TODO comments
    "NPY",  # Some numpy-specific things
]
